Reorganized project.
Everything is now in a namespace under a GDCC namespace.
Interface headers are now stored separately from the source.
When headers include other headers, it must always be by a relative path.
Relative paths should always go via the include root. That is, prefixed with .. even when including a header in the same directory.