CPP: Restructured pragma handling to separate state and parsing.
Pragma data is no longer modular, but there was no way to access the modular data without awkward dynamic typing kludges, either of the container or the data. There was never a compelling reason to modularize the data, anyway. Avoiding a few vectors in what amounts to a singleton should not be a design priority.