ACC: Removed extended type specifiers except fixed and float.
Extensions should only be added with actual use cases and after careful consideration and design. Not simply because the compiler happens to be able to support them. In particular, long, short, and unsigned introduced multi-token type specifiers to a language that was otherwise without them. Not an addition to be made lightly.
The float type specifier is left in because there are enough instances of the s15.16 fixed-point type having insufficient range for intermediary results in calculations.