ACC: Split object and function/script parsing from Parse/Declaration.cpp.
This split, a continuation of the special and print splits, is for two primary reasons. First, Parse/Declaration.cpp was getting too big. Secondly, in ACS, these declarations are all distinct grammatical entities.
This also marks the style switch for section header comments from "Global" to "Extern" and "Variable" to "Object" to be more in keeping with C++ terminology.