ACC: Removed the self-import pass.
The self-import is a way to emulate acc's ability to forward call functions, however, it also introduces the ability to forward reference objects which acc does not do, and it has been the source of too many bugs. A solution for forward-calling functions will need to be added in C-style, ie silently generating outrageously broken code.