Platform: Added MustEmitObject function.
As it turns out, undefined and unused map array objects do still have a use: Taking up space in the map array space. So they need to be kept around to properly generate AIMP. Of course, everything would work have been fine if AIMP didn't specify the size of the array in the first place.