libc: Moved formatting code out of libGDCC into libc.
The formatting logic is very specific to C and not likely to be generalized. And even if it is, any language's standard library that has sufficiently C-like formatting rules is likely to depend on libc, anyway.
While moving the code, some changes were made. First, parameters and members were renamed to use reserved names. Second, the format buffer is now extern so it can be reused in scanf code.