libc: Rewrote external objects to have getter functions.
This allows use of libc as a separate binary, as it is not necessary for using code to know the exact address of static objects at compile time. The cost is slightly slower access time, but it should not be a problem. For situations where it might be, the define __GDCC__DirectObject will restore the original behavior for such macros.