CC: Add structure properties
This is a highly experimental feature which is likely to see syntactic
and semantic evolution over time. It is intended as a means to simplify
accessing members of in-engine objects (Actors for ZDACS, Thinkers for
DGE) by allowing them to be accessed using normal structure syntax. It
also allows for using something not unlike C++ member functions.