IR: Move Code size to Arg and update SR
This is the initial step in implementing this change. Moving size into
Arg will allow proper support for mixed-size ops, rather than being
inferred from the code. In the long term, this means potential
performance improvements from generating mixed size arithmetic ops. In
the short term, it is needed for the Tr instructions for cleaning up
the conversion codegen.