Platform: Move in Addr/CallType and rename to Target
This resolves the circular dependency between Platform and IR.

Target is a more descriptive name for the module and now was a good
opportunity to rename. There was a wide overlap of files touched and
both changes are largely compiler-checked syntactic changes.

This also changes the __GDCC_Target predefined macros to __GDCC_Engine
and the --bc-target/format options to --target-engine/format.