IR: Reorganize Jfar/Jset into Jfar_Pro/Set/Sta
It never made sense for start and propagate to be the same instruction, since
they perform totally separate actions, aside from the branch at the end.

Jfar_Set not takes addr,jmpbuf instead of jmpbuf,addr to be consistent with the
other two.

Jfar_Sta does not take the drop arg. The drop argument in Jfar_Pro is because
the drop is conditional. In Jfar_Sta it is unconditional. There is also no
obvious reason to have unneeded data on the stack when starting a far jump.