BC: Implement Tr code
This effectively moves the conversion codegen from SR to BC, and most of the
code in the former has been removed. The fixed-fixed conversion is still done
in SR, to simplify the initial handling in the bytecode backends. Boolean
conversion is also done in SR, because boolean is not an IR code type and the
conversion is trivially inlined.

Subword Tr needs to be completed in BC::DGE, but given the above, there is no
resulting feature regression.