BC::ZDACS: Reduce swaps in Div and Call/Retn codegen
There's a missed opportunity to avoid a swap in Call for single-word
dst and non-stack src, but removing the (checked) assumption on stack
src would be too much complexity to be worthwhile at this time. Plus,
the higher level codegen is not set up to make of use of such a change.