SR: Push address before data where needed
The semantics for stack ordering of moves is now target-dependent.

Aut stores are still in data-addr order to accomodate adding the stack
pointer in, and so still require a swap. I think the correct change to
address this is to avoid use of Aut IR args for targets where it is not
natively supported.